+++ src/runtime/rwindow.r
@@ -4725,7 +4725,7 @@ char * watt(wbp w, char *s)
 
    if (strchr(s, '=') == NULL) {
       if (is:integer(throw)) {
-	 sprintf(foo, "%ld", IntVal(throw));
+	 sprintf(foo, "%ld", (long)(IntVal(throw)));
 	 }
       else if (!is:string(throw)) return NULL;
       return strdup(foo);
